Update according to the comments in PR
